# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Fitness platform REST API — training programs, nutrition diary, Google Calendar integration.",
        "title": "FitProfi API",
        "contact": {},
        "version": "0.1.0"
    },
    "host": "localhost:8086",
    "basePath": "/api/v1",
    "paths": {
...
        "/profiles/user": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns the fitness profile of the authenticated user.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Profiles"
                ],
                "summary": "Get user profile",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_msskobelina_fit-profi_internal_domain_model.UserProfile"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_msskobelina_fit-profi_internal_delivery_controller.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_msskobelina_fit-profi_internal_delivery_controller.ErrorResponse"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Replaces the fitness profile for the authenticated user.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Profiles"
                ],
                "summary": "Update user profile",
                "parameters": [
                    {
                        "description": "Updated profile data",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_delivery_controller_profiles.UpdateUserProfileRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_msskobelina_fit-profi_internal_domain_model.UserProfile"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_msskobelina_fit-profi_internal_delivery_controller.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_msskobelina_fit-profi_internal_delivery_controller.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Creates a fitness profile for the authenticated user.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Profiles"
                ],
                "summary": "Create user profile",
                "parameters": [
                    {
                        "description": "User profile data",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_delivery_controller_profiles.CreateUserProfileRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_msskobelina_fit-profi_internal_domain_model.UserProfile"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_msskobelina_fit-profi_internal_delivery_controller.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_msskobelina_fit-profi_internal_delivery_controller.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/programs": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns all training programs belonging to the authenticated user.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Programs"
                ],
                "summary": "List training programs",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/github_com_msskobelina_fit-profi_internal_domain_model.TrainingProgram"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_msskobelina_fit-profi_internal_delivery_controller.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_msskobelina_fit-profi_internal_delivery_controller.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Creates a new training program with optional days and exercises.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Programs"
                ],
                "summary": "Create training program",
                "parameters": [
                    {
                        "description": "Program data",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_delivery_controller_programs.CreateProgramRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_msskobelina_fit-profi_internal_domain_model.TrainingProgram"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_msskobelina_fit-profi_internal_delivery_controller.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_msskobelina_fit-profi_internal_delivery_controller.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/programs/progress": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Records a completed set of an exercise for the authenticated user.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Programs"
                ],
                "summary": "Track exercise progress",
                "parameters": [
                    {
                        "description": "Progress data",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_delivery_controller_programs.TrackProgressRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_msskobelina_fit-profi_internal_domain_model.ExerciseProgress"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_msskobelina_fit-profi_internal_delivery_controller.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_msskobelina_fit-profi_internal_delivery_controller.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/programs/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns a training program by ID.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Programs"
                ],
                "summary": "Get training program",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Program ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_msskobelina_fit-profi_internal_domain_model.TrainingProgram"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_msskobelina_fit-profi_internal_delivery_controller.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_msskobelina_fit-profi_internal_delivery_controller.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Deletes a training program owned by the authenticated user.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Programs"
                ],
                "summary": "Delete training program",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Program ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_msskobelina_fit-profi_internal_delivery_controller.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_msskobelina_fit-profi_internal_delivery_controller.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "github_com_msskobelina_fit-profi_internal_delivery_controller.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "validation error message"
                }
            }
        },
...
        "github_com_msskobelina_fit-profi_internal_domain_model.ExerciseProgress": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "exerciseId": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "notes": {
                    "type": "string"
                },
                "reps": {
                    "type": "integer"
                },
                "sets": {
                    "type": "integer"
                },
                "updatedAt": {
                    "type": "string"
                },
                "userId": {
                    "type": "integer"
                },
                "weightKg": {
                    "type": "integer"
                }
            }
        },
        "github_com_msskobelina_fit-profi_internal_domain_model.Goal": {
            "type": "string",
            "enum": [
                "lose_weight",
                "gain_weight",
                "rehab",
                "keep_fit",
                "competition"
            ],
            "x-enum-varnames": [
                "GoalLoseWeight",
                "GoalGainWeight",
                "GoalRehab",
                "GoalKeepFit",
                "GoalCompetition"
            ]
        },
        "github_com_msskobelina_fit-profi_internal_domain_model.ProgramDay": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "dayNumber": {
                    "type": "integer"
                },
                "exercises": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_msskobelina_fit-profi_internal_domain_model.ProgramExercise"
                    }
                },
                "id": {
                    "type": "integer"
                },
                "programId": {
                    "type": "integer"
                },
                "title": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "github_com_msskobelina_fit-profi_internal_domain_model.ProgramExercise": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "dayId": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "notes": {
                    "type": "string"
                },
                "reps": {
                    "type": "integer"
                },
                "sets": {
                    "type": "integer"
                },
                "updatedAt": {
                    "type": "string"
                },
                "weightKg": {
                    "type": "integer"
                }
            }
        },
        "github_com_msskobelina_fit-profi_internal_domain_model.TrainingProgram": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "days": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_msskobelina_fit-profi_internal_domain_model.ProgramDay"
                    }
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "title": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                },
                "userId": {
                    "type": "integer"
                }
            }
        },
        "github_com_msskobelina_fit-profi_internal_domain_model.UserProfile": {
            "type": "object",
            "properties": {
                "age": {
                    "type": "integer"
                },
                "createdAt": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "fullName": {
                    "type": "string"
                },
                "goal": {
                    "$ref": "#/definitions/github_com_msskobelina_fit-profi_internal_domain_model.Goal"
                },
                "id": {
                    "type": "integer"
                },
                "updatedAt": {
                    "type": "string"
                },
                "userId": {
                    "type": "integer"
                },
                "weightKg": {
                    "type": "number"
                }
            }
        },
...
        "internal_delivery_controller_profiles.CreateUserProfileRequest": {
            "type": "object",
            "required": [
                "age",
                "fullName",
                "goal",
                "weightKg"
            ],
            "properties": {
                "age": {
                    "type": "integer",
                    "example": 25
                },
                "description": {
                    "type": "string",
                    "example": "I want to stay healthy"
                },
                "fullName": {
                    "type": "string",
                    "example": "John Doe"
                },
                "goal": {
                    "type": "string",
                    "enum": [
                        "lose_weight",
                        "gain_weight",
                        "rehab",
                        "keep_fit",
                        "competition"
                    ],
                    "example": "keep_fit"
                },
                "weightKg": {
                    "type": "number",
                    "example": 75.5
                }
            }
        },
...
        "internal_delivery_controller_profiles.UpdateUserProfileRequest": {
            "type": "object",
            "required": [
                "age",
                "fullName",
                "goal",
                "weightKg"
            ],
            "properties": {
                "age": {
                    "type": "integer",
                    "example": 26
                },
                "description": {
                    "type": "string",
                    "example": "Updated description"
                },
                "fullName": {
                    "type": "string",
                    "example": "John Doe"
                },
                "goal": {
                    "type": "string",
                    "enum": [
                        "lose_weight",
                        "gain_weight",
                        "rehab",
                        "keep_fit",
                        "competition"
                    ],
                    "example": "lose_weight"
                },
                "weightKg": {
                    "type": "number",
                    "example": 74
                }
            }
        },
        "internal_delivery_controller_programs.CreateProgramRequest": {
            "type": "object",
            "required": [
                "title"
            ],
            "properties": {
                "days": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_msskobelina_fit-profi_internal_domain_model.ProgramDay"
                    }
                },
                "description": {
                    "type": "string",
                    "example": "3-day full-body routine"
                },
                "title": {
                    "type": "string",
                    "example": "Beginner Full Body"
                }
            }
        },
        "internal_delivery_controller_programs.TrackProgressRequest": {
            "type": "object",
            "required": [
                "exerciseId",
                "reps",
                "sets"
            ],
            "properties": {
                "exerciseId": {
                    "type": "integer",
                    "example": 5
                },
                "notes": {
                    "type": "string",
                    "example": "Felt strong today"
                },
                "reps": {
                    "type": "integer",
                    "example": 12
                },
                "sets": {
                    "type": "integer",
                    "example": 3
                },
                "weightKg": {
                    "type": "integer",
                    "minimum": 0,
                    "example": 60
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Enter: Bearer \u003ctoken\u003e",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}